+++ src/sites/dashboard/index.jsx
@@ -1,7 +1,19 @@
-import { h, render, Component } from 'preact';
-import { translate } from 'services';
-import { IframeCard } from 'components';
-import { connect } from 'core';
-import MainContent from './components/MainContent/index'
+/*
+This is a modified version of the file https://github.com/converge/instapy-dashboard/blob/master/src/components/MainContent/index.js
+*/
+import {h} from 'preact'
+import React, { Component } from 'react'
+import { Switch, Route } from 'react-router-dom'
+import Statistics from './components/statistics'
+import UserStatistics from './components/user_statistics'
 
-export default MainContent;
+export default class MainContent extends Component {
+  render({ match }) {
+    return (
+        <Switch>
+          <Route path={`${match.url}/`} exact component={Statistics} />
+          <Route path={`${match.url}/userStatistics/:id`} component={UserStatistics} />
+        </Switch>
+    )
+  }
+}
